Add test suite, coverage gate, MCP e2e test, and static analysis
- 69 GoogleTest cases over model, snapshot builder, sequencer timing,
  automation curves, generators, renderer, control dispatcher, socket
  framing, and live AU hosting; no audio device needed
- scripts/coverage.sh: llvm-cov with an enforced 80% line gate (at 85.5%)
- scripts/e2e_mcp.py: drives the app through the real MCP bridge and
  asserts on rendered audio
- scripts/static-analysis.sh + .clang-tidy: clang-tidy and cppcheck, all
  findings fixed; sonar-project.properties documents the licensing paths
- Split ControlDispatcher out of ControlServer so the API is unit-testable
- AppServices can skip opening an audio device (tests, headless tools)

+++ src/engine/AudioClipCache.cpp
@@ -71,7 +71,7 @@ std::shared_ptr<const juce::AudioBuffer<float>> AudioClipCache::getStretched (co
     stretcher.study (inPtrs, (size_t) numIn, true);
     stretcher.process (inPtrs, (size_t) numIn, true);
 
-    const int avail = (int) stretcher.available();
+    const int avail = stretcher.available();
     auto out = std::make_shared<juce::AudioBuffer<float>> (2, juce::jmax (1, avail));
     float* outPtrs[2] = { out->getWritePointer (0), out->getWritePointer (1) };
     stretcher.retrieve (outPtrs, (size_t) juce::jmax (0, avail));
